pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.2",
    "openai>=1.101.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "streamlit>=1.48.1",
]
